requirements.txt
cupy-cuda12x==12.3.0
graph-transformer-pytorch==0.0.3
rotary-embedding-torch==0.2.1
numpy==1.23.4
pandas==1.5.1
scikit-learn==1.1.3
torch==2.3.0
xgboost==1.6.2